[Editor] Refactor Editor => TextEditor
